requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
PyMuPDF==1.18.17
google-generativeai==0.3.1
python-dotenv==0.19.0
sqlalchemy==1.4.23
pydantic==1.8.2